fix registry & vpn stack deploy tracking
